+++ src/themeConfig.js
@@ -1,0 +1,15 @@
+import { createMuiTheme } from '@material-ui/core';
+import indigo from '@material-ui/core/colors/indigo';
+import deepOrange from '@material-ui/core/colors/deepOrange';
+
+const theme = createMuiTheme({
+  palette: {
+    primary: indigo,
+    secondary: deepOrange
+  },
+  typography: {
+    useNextVariants: true
+  }
+});
+
+export default theme;
